Made changes to variable type of the healthBar. The game implementation is left in the Play Panel.

+++ Dice Game/src/cse360project/Batman.java
@@ -29,11 +29,11 @@ public class Batman extends Character{
 		return weapon3;
 	}
 	
-	void updateHealth(double health){
+	void updateHealth(int health){
 		healthBar = health;
 	}
 	
-	double getHealth(){
+	int getHealth(){
 		return healthBar;
 	}
